refactor : 도메인 패키지 네이밍 엔티티로 변경
