+++ src/main/java/cart/product/dao/ProductDao.java
@@ -10,7 +10,7 @@ import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import cart.product.domain.Product;
+import cart.product.entity.Product;
 
 @Component
 public class ProductDao {
