+++ src/main/java/cart/product/entity/Product.java
@@ -1,4 +1,4 @@
-package cart.product.domain;
+package cart.product.entity;
 
 import java.time.LocalDateTime;
